refactor: users page
- moved internal components out of page component file
- made user forms responsive
