+++ src/components/pages/users/user-form/index.js
@@ -1,0 +1,1 @@
+export * from './user-form';
