+++ src/components/pages/users/user-list/index.js
@@ -1,0 +1,1 @@
+export { default } from './user-list.jsx';
