+++ src/components/pages/users/user-list/user-list.module.scss
@@ -1,0 +1,64 @@
+@import "styles/_variables";
+
+.avatar {
+  width: 24px;
+  height: 24px;
+}
+
+.user-list {
+  display: block;
+}
+
+.user-list-wrapper {
+  overflow-y: auto;
+}
+
+.user-list-header {
+  display: grid;
+  grid-template-columns: 24px auto 75px 75px 75px 125px 125px 24px;
+  grid-gap: 15px;
+  padding: 15px;
+}
+
+.user-list-header div {
+  font-size: 11px;
+  font-weight: 700;
+  line-height: 14px;
+  text-transform: uppercase;
+  user-select: none;
+}
+
+.user-list-item {
+  display: grid;
+  grid-template-columns: 24px auto 75px 75px 75px 125px 125px 24px;
+  grid-gap: 15px;
+  box-shadow: inset 0px 0.5px 0px rgba(143, 144, 166, 0.5);
+
+  padding: 19px 15px;
+  text-align: left;
+  background: #fff;
+}
+
+.user-list-item div {
+  display: flex;
+  align-items: center;
+  color: #627386;
+  font-size: 14px;
+
+  &:nth-child(2) {
+    color: black;
+  }
+
+  svg {
+    color: #b0b1c3;
+    width: 22px;
+    height: 22px;
+  }
+}
+
+@media only screen and (max-width: 992px) {
+  .user-list-header,
+  .user-list-item {
+    min-width: 800px;
+  }
+}
