+++ src/components/shared/form/input/input.module.scss
@@ -1,3 +1,6 @@
+@import "styles/_variables";
+@import "styles/mixins";
+
 .input {
   &-wrapper {
     display: flex;
@@ -14,6 +17,9 @@
   font-size: 14px;
   line-height: 22px;
   color: #343a41;
+  @include sm-down {
+    width: 100% !important;
+  }
   &-with-icon {
     padding: 6px 12px 6px 34px;
   }
