+++ src/components/shared/modal/modal.module.scss
@@ -1,3 +1,6 @@
+@import "styles/_variables";
+@import "styles/mixins";
+
 .overlay,
 .modal-wrapper {
   position: fixed;
@@ -25,6 +28,10 @@
   background-color: #fff;
   border-radius: 6px;
   padding: 30px;
+  @include sm-down {
+    padding: 15px;
+    width: 96vw;
+  }
 }
 
 .modal-header {
